Test added to prove absence of stackoverflow bug
There's a bug in v2.x that causes a stackoverflow in case a type with a
cyclic dependency that is registered using RegisterOpenGeneric is
resolved. In v3 the bug does not exists, and this test is here to prove
that it isn't and to make sure it will never return. Be gone you ugly
bug!