Inconsistent behavior fixed.
The WCF integration library behaved differently when Verify() was
called. In case a WCF service was marked with [ServiceBehavior(Single)]
and decorated with a transient decorator, the application WCF service
would fail at start-up, only when Verify() wasn't called. It's very
common for developers to only call Verify() in DEBUG mode, so this could
let the WCF service only fail when deployed to test or production.
(Related to #220)