Performance optimization
Performance optimization in internal Helpers.ToFriendlyName(Type)
method. The optimization prevents any memory allocations when this
method is called on a non-generic type. Although this method is only
used when Simple Injector throws an exception, a common change will call
this method in a tight loop, which can cause more allocations than
needed.