Improvement in AspNetCore integration package.
CrossWire added a IServiceScope registration in case the cross-wired
service was either Transient or Scoped. This made it very hard for the
user to register IServiceScope himself in case there were no cross-wired
registrations. CrossWire now uses unregistered type resolution to
register IServiceScope at the moment it is requested for the first time.